Code review fix. It is need to keep interface and implementation together.
